requirements.txt
# Core Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[postgresql]==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
bcrypt==4.1.2

# LangSmith Integration
langsmith==0.0.69
langchain==0.0.350

# Background Jobs & Caching
celery==5.3.4
redis==5.0.1
hiredis==2.2.3

# HTTP Client & Utilities
httpx==0.25.2
aiofiles==23.2.1

# Testing Framework Dependencies
jsonschema==4.20.0
textblob==0.17.1

# Statistical Analysis for A/B Testing
numpy==1.24.3
scipy==1.11.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Date/time utilities
python-dateutil==2.8.2

# Environment management
python-dotenv==1.0.0

# Monitoring and observability (for production)
prometheus-client==0.19.0

# Rate limiting utilities
slowapi==0.1.9

# Data export formats
openpyxl==3.1.2 